RibbonLib, changes
Added:
- RBNLoopCreateRibbonStatements for creating MapBasic statements to
create the full Ribbon. This will create the statements for recreating
most of the ribbon using the RibbonLib. However some control types
aren't yet supported and therefor ignore, such as the Gallery. Also it's
not possible to extract all informations from the Ribbon, certain
information will set to a generic value, such as the ToolTips, the Menu
Command ID and the Image. Some of the controls have been assigned the
matching image, but not all
- RBNControlGetSizeThs(...) As Integer
- RBNSelectTab(...)
- Added ParentControlType to the Elements arrays - to be able to remove
controls from SplitButton and DropDButtons
- Fixed RBNClearCustomRibbonElements to better support SplitButton and
DropDButton Controls